refactor(frontend): split `useOptionsFormDataStore` into multliple composables
Split the store into smaller, concern-specific composables for improved clarity and maintainability.
Also rename query composables to better reflect their purpose.
